+++ types/images.d.ts
@@ -1,0 +1,7 @@
+declare module '*.inline.svg' {
+  import React = require('react')
+
+  export const ReactComponent: React.FC<React.SVGProps<SVGSVGElement>>
+  const src: string
+  export default ReactComponent
+}